Moved checkBookingLimits to backend only code
