+++ packages/lib/index.ts
@@ -2,4 +2,3 @@ export { default as classNames } from "./classNames";
 export { default as isPrismaObj, isPrismaObjOrUndefined } from "./isPrismaObj";
 export * from "./isRecurringEvent";
 export * from "./isBookingLimits";
-export * from "./checkBookingLimits";
